refactor(sitTypes): move into /integration/sit.ts
