+++ components/configuration/ConfigBar.tsx
@@ -14,7 +14,7 @@ import React, { useMemo } from "react";
 import { useUser } from "@auth0/nextjs-auth0/client";
 import MobileConfigUpdateBar from "./MobileConfigUpdateBar";
 import { arraysAreEqual } from "../../utils/arrayUtils";
-import { SitClass } from "../../types/sitTypes";
+import { SitClass } from "../../types/integration/sit";
 import { DateTime } from "luxon";
 import { ClassConfig, NotificationsConfig, UserConfig } from "../../types/rezervoTypes";
 import { classConfigRecurrentId, sitClassRecurrentId } from "../../lib/iBooking";
