+++ components/modals/Agenda/Agenda.tsx
@@ -2,7 +2,7 @@ import { Box, Typography, useTheme } from "@mui/material";
 import AgendaClassItem, { AgendaClass } from "./AgendaClassItem";
 import React from "react";
 import { classConfigRecurrentId } from "../../../lib/iBooking";
-import { SitClass } from "../../../types/sitTypes";
+import { SitClass } from "../../../types/integration/sit";
 import { ClassConfig } from "../../../types/rezervoTypes";
 import { WEEKDAY_NUMBER_TO_NAME } from "../../../utils/timeUtils";
