+++ components/modals/ClassInfo/ClassInfoModal.tsx
@@ -3,7 +3,7 @@ import ClassInfo from "./ClassInfo";
 import { sitClassRecurrentId } from "../../../lib/iBooking";
 import { AllConfigsIndex, ClassPopularity, ClassPopularityIndex } from "../../../types/rezervoTypes";
 import React, { Dispatch, SetStateAction } from "react";
-import { SitClass } from "../../../types/sitTypes";
+import { SitClass } from "../../../types/integration/sit";
 
 const ClassInfoModal = ({
     classInfoClass,
