+++ components/schedule/class/ClassPopularityMeter.tsx
@@ -1,7 +1,7 @@
 import SpeedRoundedIcon from "@mui/icons-material/SpeedRounded";
 import QuestionMarkRoundedIcon from "@mui/icons-material/QuestionMarkRounded";
 import { ClassPopularity, StatusColors } from "../../../types/rezervoTypes";
-import { SitClass } from "../../../types/sitTypes";
+import { SitClass } from "../../../types/integration/sit";
 import { isClassInThePast } from "../../../lib/iBooking";
 import { stringifyClassPopularity, determineClassPopularity } from "../../../lib/popularity";
 import RippleBadge from "../../utils/RippleBadge";
